Added custom title pages.
Added variable: "basetitle" to the config file which sets the title on
index, login and registration pages as well as the admin pages. I then
moved the code that connects to the SMTP server to above where the user
is created in registration. I did this because if the connection fails
then it wont continue and create a user that wont be able get an
activation link. I also adjusted some indentions to make it easier to
read on github.com